requirements.txt
aiogram
pymongo
dnspython
python-dotenv
flask
uvicorn
gunicorn